#131 Fix: 바이낸스 api 응답 자료형 변경

+++ src/main/java/com/memesphere/domain/collection/dto/response/CollectionGridPreviewResponse.java
@@ -22,13 +22,13 @@ public class CollectionGridPreviewResponse {
     @Schema(description = "밈코인 image", example = "https://bucketname.s3.region.amazonaws.com/image1.jpg")
     String image;
     @Schema(description = "차트 데이터의 price", example = "2000")
-    BigDecimal currentPrice;
+    Double currentPrice;
     @Schema(description = "차트 데이터의 high_price", example = "2500")
-    BigDecimal highPrice;
+    Double highPrice;
     @Schema(description = "차트 데이터의 low_price", example = "1500")
-    BigDecimal lowPrice;
+    Double lowPrice;
     @Schema(description = "차트 데이터의 price_change", example = "500")
-    BigDecimal priceChange;
+    Double priceChange;
     @Schema(description = "차트 데이터의 price_change_rate", example = "2.4")
-    BigDecimal priceChangeRate;
+    Double priceChangeRate;
 }
